requirements.txt
asgiref==3.8.1
# Django==5.0.7
django-ckeditor-5==0.2.13
django-debug-toolbar==4.4.6
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
